# chaos-api/src/main/resources/static/doc/openapi.json
{
  "components": {
    "schemas": {
...
      "SysPermissionUpDTO_request": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "权限id",
            "type": "number",
            "example": "0"
          },
          "permissionName": {
            "format": "string",
            "description": "权限名称",
            "type": "string"
          }
        }
      },
      "SysPermissionVO_response": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "权限id",
            "type": "number",
            "example": "0"
          },
          "permissionName": {
            "format": "string",
            "description": "权限名称",
            "type": "string",
            "example": ""
          }
        }
      },
      "SysPostDTO_request": {
        "properties": {
          "postCode": {
            "format": "string",
            "description": "岗位编码",
            "type": "string"
          },
          "postName": {
            "format": "string",
            "description": "岗位名称",
            "type": "string"
          },
          "sort": {
            "format": "int32",
            "description": "显示顺序",
            "type": "integer",
            "example": "0"
          },
          "status": {
            "format": "int32",
            "description": "状态（0正常 1停用）",
            "type": "integer",
            "example": "0"
          }
        }
      },
...
      "SysPostUpDTO_request": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "岗位id",
            "type": "number",
            "example": "0"
          },
          "postCode": {
            "format": "string",
            "description": "岗位编码",
            "type": "string"
          },
          "postName": {
            "format": "string",
            "description": "岗位名称",
            "type": "string"
          },
          "sort": {
            "format": "int32",
            "description": "显示顺序",
            "type": "integer",
            "example": "0"
          },
          "status": {
            "format": "int32",
            "description": "状态（0正常 1停用）",
            "type": "integer",
            "example": "0"
          }
        }
      },
...
      "CommonResultSysPermissionVO_response": {
        "properties": {
          "success": {
            "format": "boolean",
            "description": "是否成功",
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "format": "string",
            "description": "响应信息",
            "type": "string",
            "example": ""
          },
          "data": {
            "description": "响应数据(object)",
            "type": "object",
            "$ref": "#/components/schemas/SysPermissionVO_response"
          },
          "code": {
            "format": "string",
            "description": "响应码",
            "type": "string",
            "example": ""
          },
          "timestamp": {
            "format": "string",
            "description": "时间戳",
            "type": "string",
            "example": ""
          },
          "traceId": {
            "format": "string",
            "description": "请求ID",
            "type": "string",
            "example": ""
          }
        }
      },
      "CommonResultSysPostVO_response": {
        "properties": {
          "success": {
            "format": "boolean",
            "description": "是否成功",
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "format": "string",
            "description": "响应信息",
            "type": "string",
            "example": ""
          },
          "data": {
            "description": "响应数据(object)",
            "type": "object",
            "$ref": "#/components/schemas/SysPostVO_response"
          },
          "code": {
            "format": "string",
            "description": "响应码",
            "type": "string",
            "example": ""
          },
          "timestamp": {
            "format": "string",
            "description": "时间戳",
            "type": "string",
            "example": ""
          },
          "traceId": {
            "format": "string",
            "description": "请求ID",
            "type": "string",
            "example": ""
          }
        }
      },
...
      "SysPermissionDTO_request": {
        "properties": {
          "permissionName": {
            "format": "string",
            "description": "权限名称",
            "type": "string"
          }
        }
      },
...
      "CommonResultBoolean_response": {
        "properties": {
          "success": {
            "format": "boolean",
            "description": "是否成功",
            "type": "boolean",
            "example": "true"
          },
          "message": {
            "format": "string",
            "description": "响应信息",
            "type": "string",
            "example": ""
          },
          "data": {
            "format": "boolean",
            "description": "响应数据",
            "type": "boolean"
          },
          "code": {
            "format": "string",
            "description": "响应码",
            "type": "string",
            "example": ""
          },
          "timestamp": {
            "format": "string",
            "description": "时间戳",
            "type": "string",
            "example": ""
          },
          "traceId": {
            "format": "string",
            "description": "请求ID",
            "type": "string",
            "example": ""
          }
        }
      },
...
      "SysPostVO_response": {
        "properties": {
          "id": {
            "format": "int64",
            "description": "岗位id",
            "type": "number",
            "example": "0"
          },
          "postCode": {
            "format": "string",
            "description": "岗位编码",
            "type": "string",
            "example": ""
          },
          "postName": {
            "format": "string",
            "description": "岗位名称",
            "type": "string",
            "example": ""
          },
          "sort": {
            "format": "int32",
            "description": "显示顺序",
            "type": "integer",
            "example": "0"
          },
          "status": {
            "format": "int32",
            "description": "状态（0正常 1停用）",
            "type": "integer",
            "example": "0"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/permission/{id}": {
      "get": {
        "summary": "根据id查询权限信息",
        "deprecated": false,
        "description": "根据id查询权限信息",
        "tags": [
          "基础模块-权限管理"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultSysPermissionVO_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": {\n    \"id\": 0,\n    \"permissionName\": \"\"\n  },\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-permission-{id}-GET",
        "parameters": [
          {
            "schema": {
              "format": "integer",
              "type": "int64"
            },
            "in": "path",
            "description": "权限id",
            "required": true,
            "name": "id"
          }
        ]
      }
    },
...
    "/api/v1/post/{ids}": {
      "delete": {
        "summary": "删除岗位信息",
        "deprecated": false,
        "description": "删除岗位信息",
        "tags": [
          "基础模块-岗位管理"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-post-{ids}-DELETE",
        "parameters": [
          {
            "schema": {
              "type": "object"
            },
            "in": "path",
            "description": "岗位id数组,[array of int64]",
            "required": true,
            "name": "ids"
          }
        ]
      }
    },
    "/api/v1/post/{id}": {
      "get": {
        "summary": "根据id查询岗位信息",
        "deprecated": false,
        "description": "根据id查询岗位信息",
        "tags": [
          "基础模块-岗位管理"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultSysPostVO_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": {\n    \"id\": 0,\n    \"postCode\": \"\",\n    \"postName\": \"\",\n    \"sort\": 0,\n    \"status\": 0\n  },\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-post-{id}-GET",
        "parameters": [
          {
            "schema": {
              "format": "integer",
              "type": "int64"
            },
            "in": "path",
            "description": "岗位id",
            "required": true,
            "name": "id"
          }
        ]
      }
    },
...
    "/api/v1/post/edit": {
      "put": {
        "summary": "更新岗位信息",
        "deprecated": false,
        "description": "更新岗位信息",
        "tags": [
          "基础模块-岗位管理"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysPostUpDTO_request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": 0,\n  \"postCode\": \"\",\n  \"postName\": \"\",\n  \"sort\": 0,\n  \"status\": 0\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-post-edit-PUT",
        "parameters": []
      }
    },
    "/api/v1/permission/{ids}": {
      "delete": {
        "summary": "删除权限信息",
        "deprecated": false,
        "description": "删除权限信息",
        "tags": [
          "基础模块-权限管理"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-permission-{ids}-DELETE",
        "parameters": [
          {
            "schema": {
              "type": "object"
            },
            "in": "path",
            "description": "权限id数组,[array of int64]",
            "required": true,
            "name": "ids"
          }
        ]
      }
    },
...
    "/api/v1/permission/edit": {
      "put": {
        "summary": "更新权限信息",
        "deprecated": false,
        "description": "更新权限信息",
        "tags": [
          "基础模块-权限管理"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysPermissionUpDTO_request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": 0,\n  \"permissionName\": \"\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-permission-edit-PUT",
        "parameters": []
      }
    },
...
    "/api/v1/post/save": {
      "post": {
        "summary": "保存岗位信息",
        "deprecated": false,
        "description": "保存岗位信息",
        "tags": [
          "基础模块-岗位管理"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysPostDTO_request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"postCode\": \"\",\n  \"postName\": \"\",\n  \"sort\": 0,\n  \"status\": 0\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-post-save-POST",
        "parameters": []
      }
    },
    "/api/v1/permission/save": {
      "post": {
        "summary": "保存权限信息",
        "deprecated": false,
        "description": "保存权限信息",
        "tags": [
          "基础模块-权限管理"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SysPermissionDTO_request"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"permissionName\": \"\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResultBoolean_response"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"success\": true,\n  \"message\": \"\",\n  \"data\": true,\n  \"code\": \"\",\n  \"timestamp\": \"\",\n  \"traceId\": \"\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "api-v1-permission-save-POST",
        "parameters": []
      }
    },
...
  },
  "info": {
    "title": "chaos接口文档",
    "version": "1.0.0"
  },
...
}